// src/app/clearing/models/mailsb.ts
import { GlobalVariables } from '../../core/models/globalvariables';

export class MailSB {

  sb_pkid: string;
  sb_mailid: string;
  sb_id: number;
  sb_from: string;
  sb_subject: string;
  sb_msg_date: string;
  sb_msg_type: string;
  sb_job_no: number;
  sb_job_date: string;
  sb_no: number;
  sb_date: string;
  sb_no2: string;
  sb_reason: string;
  row_colour: string;

  rec_mode: string;
  rec_category: string;
  _globalvariables: GlobalVariables;
}